clear experiments
