--- experiments/__init__.py
@@ -1,1 +1,0 @@
-import src.gp.dataset.accessor   # noqa: F401
