--- experiments/reg_toy_x.txt
@@ -1,150 +1,0 @@
--1.000000000000000000e+00
--9.328859060402684422e-01
--8.657718120805368844e-01
--7.986577181208054377e-01
--7.315436241610737689e-01
--6.644295302013423221e-01
--5.973154362416107643e-01
--5.302013422818792066e-01
--4.630872483221476488e-01
--3.959731543624160910e-01
--3.288590604026845332e-01
--2.617449664429529754e-01
--1.946308724832215287e-01
--1.275167785234899709e-01
--6.040268456375841311e-02
-6.711409395973255698e-03
-7.382550335570470246e-02
-1.409395973154361492e-01
-2.080536912751678180e-01
-2.751677852348992648e-01
-3.422818791946309336e-01
-4.093959731543623803e-01
-4.765100671140940491e-01
-5.436241610738254959e-01
-6.107382550335569427e-01
-6.778523489932886115e-01
-7.449664429530200582e-01
-8.120805369127517270e-01
-8.791946308724831738e-01
-9.463087248322148426e-01
-1.013422818791946511e+00
-1.080536912751677736e+00
-1.147651006711409405e+00
-1.214765100671141074e+00
-1.281879194630872298e+00
-1.348993288590603967e+00
-1.416107382550335636e+00
-1.483221476510067305e+00
-1.550335570469798530e+00
-1.617449664429530198e+00
-1.684563758389261867e+00
-1.751677852348993092e+00
-1.818791946308724761e+00
-1.885906040268456429e+00
-1.953020134228188098e+00
-2.020134228187919323e+00
-2.087248322147650992e+00
-2.154362416107382661e+00
-2.221476510067113885e+00
-2.288590604026845554e+00
-3.355704697986577223e+00
-3.422818791946308892e+00
-3.489932885906040116e+00
-3.557046979865771785e+00
-3.624161073825503454e+00
-3.691275167785235123e+00
-3.758389261744966348e+00
-3.825503355704698016e+00
-3.892617449664429685e+00
-3.959731543624160910e+00
-4.026845637583893023e+00
-4.093959731543623803e+00
-4.161073825503355472e+00
-4.228187919463087141e+00
-4.295302013422818810e+00
-4.362416107382550479e+00
-4.429530201342282147e+00
-4.496644295302013816e+00
-4.563758389261744597e+00
-4.630872483221476266e+00
-4.697986577181207934e+00
-4.765100671140939603e+00
-4.832214765100671272e+00
-4.899328859060402941e+00
-4.966442953020134610e+00
-5.033557046979865390e+00
-5.100671140939597059e+00
-5.167785234899328728e+00
-5.234899328859060397e+00
-5.302013422818792066e+00
-5.369127516778523734e+00
-5.436241610738255403e+00
-5.503355704697986184e+00
-5.570469798657717853e+00
-5.637583892617449521e+00
-5.704697986577181190e+00
-5.771812080536912859e+00
-5.838926174496644528e+00
-5.906040268456376197e+00
-5.973154362416106977e+00
-6.040268456375838646e+00
-6.107382550335570315e+00
-6.174496644295301984e+00
-6.241610738255033652e+00
-6.308724832214765321e+00
-6.375838926174496990e+00
-6.442953020134227771e+00
-6.510067114093959439e+00
-6.577181208053691108e+00
-6.644295302013422777e+00
-7.711409395973154446e+00
-7.778523489932886115e+00
-7.845637583892617783e+00
-7.912751677852349452e+00
-7.979865771812080233e+00
-8.046979865771811902e+00
-8.114093959731544459e+00
-8.181208053691275239e+00
-8.248322147651006020e+00
-8.315436241610738577e+00
-8.382550335570471134e+00
-8.449664429530201915e+00
-8.516778523489932695e+00
-8.583892617449663476e+00
-8.651006711409396033e+00
-8.718120805369128590e+00
-8.785234899328859370e+00
-8.852348993288590151e+00
-8.919463087248320932e+00
-8.986577181208053489e+00
-9.053691275167786046e+00
-9.120805369127516826e+00
-9.187919463087247607e+00
-9.255033557046980164e+00
-9.322147651006710944e+00
-9.389261744966443501e+00
-9.456375838926174282e+00
-9.523489932885906839e+00
-9.590604026845637620e+00
-9.657718120805368400e+00
-9.724832214765100957e+00
-9.791946308724831738e+00
-9.859060402684564295e+00
-9.926174496644295075e+00
-9.993288590604027632e+00
-1.006040268456375841e+01
-1.012751677852348919e+01
-1.019463087248322175e+01
-1.026174496644295253e+01
-1.032885906040268509e+01
-1.039597315436241587e+01
-1.046308724832214843e+01
-1.053020134228187921e+01
-1.059731543624160999e+01
-1.066442953020134254e+01
-1.073154362416107332e+01
-1.079865771812080588e+01
-1.086577181208053666e+01
-1.093288590604026922e+01
-1.100000000000000000e+01
